Inserindo validações no CRUD da entidade Categoria

+++ ecommerce/src/main/java/org/serratec/ecommerce/controllers/CategoriaController.java
@@ -2,7 +2,10 @@ package org.serratec.ecommerce.controllers;
 
 import java.util.List;
 
+import javax.validation.Valid;
+
 import org.serratec.ecommerce.entities.Categoria;
+import org.serratec.ecommerce.exceptions.NoSuchElementFoundException;
 import org.serratec.ecommerce.services.CategoriaService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -21,33 +24,47 @@ import org.springframework.web.bind.annotation.RestController;
 public class CategoriaController {
 	@Autowired
 	CategoriaService categoriaService;
-	
+
 	@GetMapping
 	public ResponseEntity<List<Categoria>> findAllCategoria() {
 		List<Categoria> categoriaList = categoriaService.findAllCategoria();
+		if (categoriaList.isEmpty())
+			throw new NoSuchElementFoundException("Não foi encontrada nenhuma categoria");
 		return new ResponseEntity<>(categoriaList, HttpStatus.OK);
 	}
-	
+
 	@GetMapping("/{id}")
 	public ResponseEntity<Categoria> findCategoriaById(@PathVariable Integer id) {
 		Categoria categoria = categoriaService.findCategoriaById(id);
+		if (categoria == null)
+			throw new NoSuchElementFoundException("Não foi encontrada categoria com o Id: " + id);
 		return new ResponseEntity<>(categoria, HttpStatus.OK);
 	}
-	
-	@PostMapping
-	public ResponseEntity<Categoria> saveCategoria(@RequestBody Categoria categoria) {
-		return new ResponseEntity<>(categoriaService.saveCategoria(categoria), HttpStatus.CREATED);
-	}
 
+	@PostMapping
+	public ResponseEntity<Categoria> saveCategoria(@Valid @RequestBody Categoria categoria) {
+		Categoria novaCategoria = categoriaService.saveCategoria(categoria);
+		if (novaCategoria == null)
+			throw new NoSuchElementFoundException(
+					"Já existe uma categoria salva com o nome " + "'" + categoria.getNomeCategoria() + "' ");
+		return new ResponseEntity<>(novaCategoria, HttpStatus.CREATED);
+	}
+ 
 	@PutMapping
-	public ResponseEntity<Categoria> updateCategoria(@RequestBody Categoria categoria) {
+	public ResponseEntity<Categoria> updateCategoria(@Valid @RequestBody Categoria categoria) {
+		if(categoria.getIdCategoria()==null) 
+			throw new NoSuchElementFoundException(
+					"Não foi informado um ID");		
 		Categoria categoriaAtualizado = categoriaService.updateCategoria(categoria);
+		if(categoriaAtualizado==null)
+			throw new NoSuchElementFoundException("Não foi encontrada categoria com o Id: " + categoria.getIdCategoria());
 		return new ResponseEntity<>(categoriaAtualizado, HttpStatus.OK);
 	}
 
-
 	@DeleteMapping("/{id}")
 	public ResponseEntity<String> deleteCategoriaById(@PathVariable Integer id) {
+		if(categoriaService.findCategoriaById(id)==null) 
+			throw new NoSuchElementFoundException("Não foi encontrada categoria com o Id: " + id);
 		categoriaService.deleteCategoriaById(id);
 		return new ResponseEntity<>("", HttpStatus.OK);
 	}
